Refactor intent-related instrumentation tests to use helper
This refactors the intent assertions in the open Signal/WhatsApp/Telegram instrumentation tests to use a safer helper method that accounts for failures during the execution of an action and releases the registered Intent listener. This prevents test pollution.
