+++ app/src/androidTest/kotlin/org/vinaygopinath/launchchat/MainActivityMenuTest.kt
@@ -1,19 +1,15 @@
 package org.vinaygopinath.launchchat
 
-import android.app.Instrumentation
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario.launch
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withText
-import org.hamcrest.Matchers.allOf
 import org.junit.Before
 import org.junit.Test
+import org.vinaygopinath.launchchat.helpers.AssertionHelper.assertIntentNavigation
 
 class MainActivityMenuTest {
 
@@ -26,18 +22,8 @@ class MainActivityMenuTest {
     fun launchesGithubPageOnAboutClick() {
         openActionBarOverflowOrOptionsMenu(ApplicationProvider.getApplicationContext())
 
-        Intents.init()
-        val expectedIntent = allOf(
-            IntentMatchers.hasAction(Intent.ACTION_VIEW),
-            IntentMatchers.hasData(Constants.GITHUB_REPO_URL)
-        )
-
-        Intents.intending(expectedIntent).respondWith(
-            Instrumentation.ActivityResult(0, null)
-        )
-
-        onView(withText("About")).perform(click())
-
-        intended(expectedIntent)
+        assertIntentNavigation(Intent.ACTION_VIEW, Constants.GITHUB_REPO_URL) {
+            onView(withText("About")).perform(click())
+        }
     }
 }
